refactor: Flatten the module structure of `evm-ext` and `genesis` crates
