+++ moved/src/move_execution/canonical.rs
@@ -221,7 +221,7 @@ pub(super) fn execute_canonical_transaction<
     let (mut changes, mut extensions) = session.finish_with_extensions()?;
     let mut logs = extensions.logs();
     logs.extend(evm_logs);
-    let evm_changes = moved_evm_ext::evm_native::extract_evm_changes(&extensions);
+    let evm_changes = moved_evm_ext::extract_evm_changes(&extensions);
     changes
         .squash(evm_changes)
         .expect("EVM changes must merge with other session changes");
